Split consistent-trees-hdf5 tests.
