+++ tests/frontends/test_consistent_trees_hdf5_4.py
@@ -1,0 +1,12 @@
+from ytree.frontends.consistent_trees_hdf5 import \
+    ConsistentTreesHDF5Arbor
+from ytree.utilities.testing import \
+    ArborTest, \
+    TempDirTest
+
+class ConsistentTreesHDF5ArborTest4(TempDirTest, ArborTest):
+    arbor_type = ConsistentTreesHDF5Arbor
+    test_filename = ["consistent_trees_hdf5/soa/forest_0.h5",
+                     "consistent_trees_hdf5/soa/forest_0.h5"]
+    num_data_files = 2
+    tree_skip = 20000
